+++ R/stats_and_visualization.R
@@ -350,3 +350,151 @@ searchRecombination <- function(phgObject = NULL,
 
     return(gametesSame[gametesDiff == 1])
 }
+
+
+# ----
+#' @title Visualize Graph Data
+#'
+#' @description
+#' Generates an interactive network plot for a given set of reference ranges
+#' and a set of taxa.
+#'
+#' @param x A \code{PHGDataSet} object
+#' @param samples Samples/taxa to include in plot
+#' @param sampleHighlight Sample path to highlight
+#' @param seqnames A sequence (e.g. chromosome) ID
+#' @param start Start position for ref ranges
+#' @param end End position for ref ranges
+#' @param colMajor Highlight path color
+#' @param colMinor Muted path color
+#' @param ... Additional parameters to pass for ref range inclusion
+#'
+#' @importFrom IRanges subsetByOverlaps
+#' @importFrom GenomicRanges GRanges
+#' @importFrom SummarizedExperiment assay
+#' @importFrom visNetwork visEdges
+#' @importFrom visNetwork visHierarchicalLayout
+#' @importFrom visNetwork visNetwork
+#'
+#' @export
+plotGraph <- function(
+    x,
+    samples = NULL,
+    sampleHighlight = NULL,
+    seqnames = NULL,
+    start = NULL,
+    end = NULL,
+    colMajor = "maroon",
+    colMinor = "lightgrey",
+    ...
+) {
+    # # Testing
+    # start <- 100
+    # end   <- 1000000
+    # seqnames <- "1"
+    # # samples <- c("Z001E0001", "Z001E0028", "Z001E0080")
+    # # samples <- NULL
+    # set.seed(123)
+    # samples <- sample(colnames(x), 100)
+    # # sampleHighlight <- c("Z001E0001")
+    # sampleHighlight <- sample(samples, 1)
+
+    # Filter by taxa and ref ranges
+    if (is.null(samples)) samples <- colnames(x)
+    hapTableMini <- x[, colnames(x) %in% samples]
+    hapTableMini <- IRanges::subsetByOverlaps(
+        hapTableMini,
+        GenomicRanges::GRanges(seqnames = seqnames, ranges = start:end)
+    )
+
+    # Get hap ID matrix
+    currentMatrix <- t(SummarizedExperiment::assay(hapTableMini))
+    currentMatrix[is.na(currentMatrix)] <- -128
+    colnames(currentMatrix) <- gsub("R", "", colnames(currentMatrix)) |>
+        as.numeric()
+
+    # Get ref range data frame
+    refRangeDataMini <- rowRanges(hapTableMini) |> as.data.frame()
+
+    # Group taxa by hap ID and ref range
+    taxaGroups <- lapply(seq_len(ncol(currentMatrix)), function(i) {
+        split(rownames(currentMatrix), currentMatrix[, i])
+    })
+
+    # Generate distinct IDs (hap ID + ref range ID)
+    hapIds     <- currentMatrix |> apply(2, unique, simplify = FALSE)
+    hapLevels  <- rep(names(hapIds), vapply(hapIds, length, integer(1))) |> as.numeric()
+    fullHapIds <- paste0(
+        lapply(hapIds, function(i) i[order(i)]) |> unlist(),
+        "_", hapLevels
+    )
+
+    # HTML tooltip processing
+    taxaToHtml <- function(x) {
+        vapply(x, function(i) {
+            paste0("<b>Taxa: </b>", paste(i, collapse = ", "), "</p>")
+        }, character(1))
+    }
+    tooltipVec <- lapply(taxaGroups, taxaToHtml) |> unlist()
+
+    refRangeHtml <- lapply(hapLevels, function(i) {
+        paste0(
+            "<p><b>Chr: </b>",
+            refRangeDataMini[i, ]$seqnames,
+            "<br>",
+            "<b>Range: </b>",
+            refRangeDataMini[i, ]$start,
+            " - ",
+            refRangeDataMini[i, ]$end,
+            "<br>"
+        )
+    }) |> unlist()
+
+    # Final graph data (nodes)
+    nodes <- data.frame(
+        id    = seq_along(fullHapIds),
+        label = fullHapIds,
+        level = hapLevels,
+        title = paste0(refRangeHtml, tooltipVec)
+    )
+
+    if (!is.null(sampleHighlight)) {
+        for (i in sampleHighlight) {
+            nodes$group <- ifelse(grepl(i, nodes$title), i, NA)
+            nodes$color <- ifelse(grepl(i, nodes$title), colMajor, colMinor)
+        }
+        nodes$title <- gsub(i, paste0("<b>", i, "</b>"), nodes$title)
+    } else {
+        nodes$color <- colMajor
+    }
+
+    # Final graph data (edges)
+    lne <- c()
+    rne <- c()
+    for (i in seq_len(ncol(currentMatrix) - 1)) {
+        ln <- paste0(currentMatrix[, i], "_", i)
+        rn <- paste0(currentMatrix[, i + 1], "_", i + 1)
+
+        cnxn <- paste0(ln, "+", rn) |> unique()
+
+        for (c in cnxn) {
+            splits <- strsplit(c, "\\+") |> unlist()
+            f <- which(fullHapIds == splits[1])
+            t <- which(fullHapIds == splits[2])
+            lne <- c(lne, f)
+            rne <- c(rne, t)
+        }
+    }
+
+    edges <- data.frame(
+        from = lne,
+        to = rne
+    )
+
+    # Return vis.js object
+    visNetwork::visNetwork(nodes, edges) |>
+        visNetwork::visEdges(arrows = "to") |>
+        visNetwork::visHierarchicalLayout(direction = "LR")
+}
+
+